+++ caso.practico.gch/src/main/java/com/gabriel/chanchay/dao/PersonaDao.java
@@ -3,6 +3,6 @@ package com.gabriel.chanchay.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.gabriel.chanchay.modelo.Persona;
 
-public interface PersonaDao extends JpaRepository<Persona, Integer>{
+public interface PersonaDao extends JpaRepository<Persona, Long>{
 
 }
